Normalize sequential verifier token dtype

+++ tests/v1/worker/gpu/test_sequential_verification.py
@@ -7,7 +7,9 @@ from vllm.v1.worker.gpu.spec_decode.rejection_sampler import (
 
 def test_pack_sequential_samples() -> None:
     target = torch.tensor([[10, 20, 30, 40], [10, 20, 30, 40]])
-    drafts = torch.tensor([[0, 10, 99, 30], [0, 10, 20, 30]])
+    drafts = torch.tensor(
+        [[0, 10, 99, 30], [0, 10, 20, 30]], dtype=torch.int32
+    )
 
     sampled, lengths = _pack_sequential_samples(target, drafts)
